P3 mühendislik: Python sürüm sözleşmesini netleştir, ruff/mypy/paket CI kapıları ekle
Sürüm sözleşmesi (review'daki tutarsızlık):
* requires-python ">=3.9" iddiası CI tarafından hiç doğrulanmıyordu ve
  ruff target-version=py310 / mypy python_version=3.11 ile çelişiyordu.
  Taban 3.10'a çekildi; 3.9 classifier'ı kaldırıldı, 3.13 eklendi.
* CI matrisi 3.10/3.11/3.12/3.13 (3.13 bloklayıcı).
* tests/test_packaging_contract.py: classifier listesi ile CI matrisi birebir
  aynı olmalı, requires-python tabanı matrisin en düşüğü olmalı, ruff/mypy
  hedefleri tabanın gerisinde kalmamalı. Sözleşme artık sessizce kayamaz.

Yeni CI kapıları:
* lint job — ruff check . + mypy hga mimari egitim
* package job — python -m build, twine check, wheel'i kurup TEMİZ dizinden
  import (paket verisi ve konsol script'i dahil). Repo kökünden kaynaklanan
  yanlış "çalışıyor" sonucunu engeller.

Lint/tip temizliği (163 ruff + 110 mypy hatası → 0):
* Gerçek bulgular: bpe_tokenizer'da tokenize döngüsünde ölü `i += l`
  referansı ve `res`/`b`/`k`/`v` değişken gölgelemeleri; conflict.py'de
  kullanılmayan `object_` bağlaması (varlık kontrolü çağrısı korundu).
* seyrek_tablo: register_buffer tensörleri sınıf düzeyinde bildirildi,
  ModuleList erişimi için tipli `_tablo()/_tablolar()` yardımcıları eklendi.
* E402 yalnız sys.path bootstrap'i zorunlu olan 17 dosyada, per-file-ignores
  ile bastırıldı; kural global olarak kapatılmadı.

456 test yeşil; davranış değişikliği yok.

+++ hga/experience/scoring.py
@@ -101,7 +101,7 @@ class Scoring:
         agrega = store.relations.olgu_agrega(subject_id, relation_id, object_id)
         if agrega is None:
             return 0.5
-        return agrega["score"]
+        return float(agrega["score"])
 
     def baglam_tutarliligi(self, subject: Entity, relation: Relation,
                            object_: Entity) -> float:
@@ -115,7 +115,7 @@ class Scoring:
 
     def bellek_destegi(self, store, subject_id: str, relation_id: str,
                        object_id: str) -> float:
-        return store.bellek_destegi(subject_id, relation_id, object_id)
+        return float(store.bellek_destegi(subject_id, relation_id, object_id))
 
     def yenilik(self, store, subject_id: str, relation_id: str,
                 object_id: str) -> float:
@@ -143,7 +143,7 @@ class Scoring:
         nb = sum(v * v for v in b.values()) ** 0.5
         if na == 0 or nb == 0:
             return 0.0
-        return pay / (na * nb)
+        return float(pay / (na * nb))
 
     def information_gain(self, store, object_: Entity) -> float:
         """Nesnenin bellek içinde ne kadar "bilgilendirici" (ayırt edici) olduğu.
